+++ test/sharp/index.js
@@ -43,3 +43,13 @@ test('png to jpeg with background color', () => {
   const output = require('../cat-transparent.png?background=#FF0000&format=jpg');
   expect(output).toMatchSnapshot();
 });
+
+test('with min and max sizes', () => {
+  const output = require('../cat-1000.jpg?min=600&max=800&steps=3');
+  expect(output).toMatchSnapshot();
+});
+
+test('with min and max sizes, and default steps', () => {
+  const output = require('../cat-1000.jpg?min=500&max=1000');
+  expect(output).toMatchSnapshot();
+});