+++ test/jimp/index.js
@@ -48,3 +48,28 @@ test('png to jpeg with background color', () => {
   const output = require('../cat-transparent.png?background=0xFF0000FF&format=jpg');
   expect(output).toMatchSnapshot();
 });
+
+test('with min and max sizes', () => {
+  const output = require('../cat-1000.jpg?min=600&max=800&steps=3');
+  expect(output).toMatchSnapshot();
+});
+
+test('with min and max sizes, and default steps', () => {
+  const output = require('../cat-1000.jpg?min=500&max=1000');
+  expect(output).toMatchSnapshot();
+});
+
+test('with min and max sizes options', () => {
+  const output = require('../cat-1000.jpg?minmax');
+  expect(output).toMatchSnapshot();
+});
+
+test('override min and max with sizes', () => {
+  const output = require('../cat-1000.jpg?minmax&sizes[]=100&sizes[]=200');
+  expect(output).toMatchSnapshot();
+});
+
+test('override min and max with size', () => {
+  const output = require('../cat-1000.jpg?minmax&size=100');
+  expect(output).toMatchSnapshot();
+});