+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
   
- - Add `min` and `max` options to automatically generate a number of images, and `steps` option to say how many images ([#31](herrstucki/responsive-loader/pull/31)).
+ - Added `min` and `max` options to automatically generate a number of images, and `steps` option to say how many images ([#31](https://github.com/herrstucki/responsive-loader/pull/31)).
 
 ## v1.0.0